add vnpay for order product and add api to update payment

+++ src/main/java/com/nashtech/cellphonesfake/view/ProductCardVm.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotBlank;
 public record ProductCardVm(
         Long id,
         @NotBlank String productName,
-        @NotBlank Double price,
+        @NotBlank Long price,
         String thumbnail
 ) {
 }
